post_service/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0
djangorestframework-simplejwt>=4.7.0,<5.0
django-celery-beat>=2.2.0,<3.0
django-celery-results>=2.2.0,<3.0
gunicorn>=20.1.0,<21.0
psycopg2-binary>=2.9.1,<3.0
pillow>=8.2.0,<9.0
drf-yasg
django-elasticsearch-dsl
elasticsearch